Fixed OSCSpace Controller
Added OSCSpace reactor to deal with OnTrigger events
