+++ Assets/OSCSpaceController.cs
@@ -19,68 +19,4 @@ public class OSCSpaceController : OSCTransmitterObjectBase
 
     [Header("Colliders to Ignore")]
     public List<Collider> ignoreColliders;
-
-    private void OnTriggerEnter(Collider other)
-    {
-        if (sendTriggerEnter && !ignoreColliders.Contains(other))
-        {
-            List<object> args = new List<object>();
-            if (includeEnterLocation)
-            {
-                Vector3 localPos = FindTriggerPoint(other);
-                args.Add(localPos.x);
-                args.Add(localPos.y);
-                args.Add(localPos.z);
-            }
-            SendOSCMessage(string.Format("{0}/{1}", oscAddress, "triggerEnter"), args.ToArray());
-        }
-
-    }
-
-    private void OnTriggerStay(Collider other)
-    {
-        if (sendTriggerStay && !ignoreColliders.Contains(other))
-        {
-            List<object> args = new List<object>();
-            if (includeLocalPosition)
-            {
-                args.Add(other.transform.localPosition.x);
-                args.Add(other.transform.localPosition.y);
-                args.Add(other.transform.localPosition.z);
-            }
-            SendOSCMessage(string.Format("{0}/{1}", oscAddress, "triggerStay"), args.ToArray());
-        }
-    }
-
-    private void OnTriggerExit(Collider other)
-    {
-        if (sendTriggerExit && !ignoreColliders.Contains(other))
-        {
-            List<object> args = new List<object>();
-            if (includeExitLocation)
-            {
-                Vector3 localPos = FindTriggerPoint(other);
-                args.Add(localPos.x);
-                args.Add(localPos.y);
-                args.Add(localPos.z);
-            }
-            SendOSCMessage(string.Format("{0}/{1}", oscAddress, "triggerExit"), args.ToArray());
-        }
-    }
-
-
-    // Research this...
-    private Vector3 FindTriggerPoint(Collider other)
-    {
-        Vector3 localPos = Vector3.zero;
-        RaycastHit hit;
-        if (Physics.Raycast(transform.position, transform.forward, out hit))
-        {
-            if (hit.transform.gameObject == other.gameObject)
-            {
-                localPos = transform.InverseTransformPoint(hit.point);
-            }
-        }
-        return localPos;
-    }
 }
